requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
numpy==1.24.3
openvino-runtime==2023.2.0
requests==2.31.0
python-multipart==0.0.6
scipy==1.11.4
pydantic==2.5.0